feat: implement JWT verification for category management; restrict delete and add operations to admin users

+++ backend/api/categorie.js
@@ -1,5 +1,7 @@
 const Categorie = require('../models/categorie');
 const Article = require('../models/article');
+const User = require('../models/user');
+const VerifyJWT = require('../middleware/verifyJWT');
 
 const router = require('express').Router();
 
@@ -12,8 +14,10 @@ router.get('/', async (req, res) => {
     }
 });
 
-router.delete('/del/:id', async (req, res) => {
+router.delete('/del/:id', VerifyJWT, async (req, res) => {
     if (!req.params.id) return res.status(400).send('ID is required');
+    const user = await User.findById(req.user.id);
+    if (!['admin'].includes(user.role)) return res.status(403).send('You do not have permission to delete a category');
     try {
         await Categorie.findByIdAndDelete(req.params.id);
         await Article.updateMany({ categories: req.params.id }, { $pull: { categories: req.params.id } });
@@ -23,8 +27,10 @@ router.delete('/del/:id', async (req, res) => {
     }
 });
 
-router.post('/add', async (req, res) => {
+router.post('/add', VerifyJWT, async (req, res) => {
     if (!req.body.name) return res.status(400).send('Name is required');
+    const user = await User.findById(req.user.id);
+    if (!['admin'].includes(user.role)) return res.status(403).send('You do not have permission to delete a category');
     try {
         await Categorie.create({ name: req.body.name });
         res.json({ message: 'Category successfully added' });
